requirements.txt
# Django基础依赖
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0

# Google OAuth认证相关
requests==2.31.0
oauthlib==3.2.2
python-dotenv==1.0.0
pyjwt==2.8.0

# 数据库相关
dj-database-url==2.3.0
psycopg2-binary==2.9.9
supabase==2.15.0

# 部署相关
gunicorn==21.2.0
whitenoise==6.6.0

# 其他工具
pytz==2023.3
Pillow==10.1.0